requirements.txt
asttokens==2.2.1
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
build==0.9.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
decorator==5.1.1
exceptiongroup==1.0.1
executing==1.2.0
flake8==6.0.0
h11==0.14.0
idna==3.4
ipython==8.14.0
jedi==0.18.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy==1.4.1
mypy-extensions==1.0.0
numpy==1.23.4
outcome==1.2.0
packaging==21.3
pandas==1.5.1
parso==0.8.3
pep517==0.13.0
pexpect==4.8.0
pickleshare==0.7.5
pip-tools==6.9.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.15.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
selenium==4.6.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.9.2
types-PyYAML==6.0.12.10
types-requests==2.31.0.1
types-urllib3==1.26.25.13
typing_extensions==4.7.0
urllib3==1.26.12
wcwidth==0.2.6
wsproto==1.2.0
